Schema adjustment, to handle duplicates
Instead of:

    var compat = {data: {a: {contents: {}, links: []}}};

And to call the contents of 'a', we would:

    return compat['data']['a']['contents'];

We are losing data if more than one 'a' element exists (e.g. in SVG, and HTML). It happened
in more than one situation. Instead, we will add ONE MORE level.

Changing in this fashion:

    var compat = {data: {html: {a: {contents: {}, links: []}}, css:{}, svg: {}, duplicates: {}}};

And the call of contents of 'a' becomes:

    return compat['data']['svg']['a']['contents'];  // SVG specific a entry
    return compat['data']['html']['a']['contents']; // HTML specific a entry

Log:
* Adding jsonselect path, sections index
* Moving link array method, renaming getData